refactor: split slang Rust binding facade
