+++ crates/slang/bindings/rust/syntax_tree.rs
@@ -1,0 +1,897 @@
+use std::{fmt, hash, iter, ops::Not, pin::Pin};
+
+use cxx::{SharedPtr, UniquePtr};
+use itertools::Either;
+
+use crate::{
+    CxxSV, LexedTokenAtOffset, LiteralBase, ParserExpectedSyntax, PreprocessorTrace,
+    PreprocessorTraceEmittedToken, PreprocessorTraceTokenProvenance, SVInt, SVLogic,
+    SourceLocation, SourceRange, SyntaxDiagnostic, SyntaxTreeBuffer, SyntaxTreeBufferIds, TimeUnit,
+    ffi,
+    syntax::{
+        SyntaxKind, TokenKind, TriviaKind,
+        cursor::SyntaxCursor,
+        iter::{SyntaxChildren, SyntaxElemPreorder, SyntaxIdxChildren, SyntaxNodePreorder},
+    },
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct SyntaxTriviaLoc {
+    pub buffer_id: u32,
+    pub start: usize,
+    pub end: usize,
+}
+
+#[derive(Clone, Copy)]
+pub struct SyntaxNode<'a> {
+    _ptr: Pin<&'a ffi::SyntaxNode>,
+}
+
+#[derive(Clone, Copy)]
+pub struct SyntaxToken<'a> {
+    _ptr: Pin<&'a ffi::SyntaxToken>,
+}
+
+#[derive(Clone)]
+pub struct SyntaxTree {
+    _ptr: SharedPtr<ffi::SyntaxTree>,
+}
+
+#[derive(Debug, Clone)]
+pub struct SyntaxTreeWithPreprocessorTrace {
+    pub tree: SyntaxTree,
+    pub preprocessor_trace: Option<PreprocessorTrace>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct SyntaxTreeOptions {
+    pub predefines: Vec<String>,
+    pub include_paths: Vec<String>,
+    pub include_buffers: Vec<SyntaxTreeBuffer>,
+    pub expand_includes: bool,
+}
+
+impl Default for SyntaxTreeOptions {
+    fn default() -> Self {
+        Self {
+            predefines: Vec::new(),
+            include_paths: Vec::new(),
+            include_buffers: Vec::new(),
+            expand_includes: true,
+        }
+    }
+}
+
+impl SyntaxTreeOptions {
+    pub fn without_include_expansion() -> Self {
+        Self { expand_includes: false, ..Self::default() }
+    }
+}
+
+#[derive(Clone, Copy)]
+pub struct SyntaxTrivia<'a> {
+    _ptr: Pin<&'a ffi::SyntaxTrivia>,
+}
+
+impl fmt::Debug for SyntaxTrivia<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SyntaxTrivia")
+            .field("kind", &self.kind())
+            .field("raw_text", &self.get_raw_text())
+            .finish()
+    }
+}
+
+pub trait ChildrenIter<It> = DoubleEndedIterator<Item = It> + ExactSizeIterator + Clone;
+
+impl<'a> SyntaxTrivia<'a> {
+    #[inline]
+    fn from_raw_ptr(_ptr: *const ffi::SyntaxTrivia) -> Option<Self> {
+        assert!(_ptr.is_null().not());
+        Some(SyntaxTrivia { _ptr: unsafe { Pin::new_unchecked(&*_ptr) } })
+    }
+
+    #[inline]
+    pub fn get_raw_text(&self) -> CxxSV<'_> {
+        self._ptr.getRawText()
+    }
+
+    #[inline]
+    pub fn kind(&self) -> TriviaKind {
+        TriviaKind::from_id(self._ptr.kind())
+    }
+
+    #[inline]
+    fn explicit_location(&self) -> Option<SourceLocation> {
+        SourceLocation::from_unique_ptr(self._ptr.getExplicitLocation())
+    }
+
+    #[inline]
+    pub fn syntax(&self) -> Option<SyntaxNode<'a>> {
+        SyntaxNode::from_raw_ptr(self._ptr.syntax())
+    }
+}
+
+impl<'a> SyntaxToken<'a> {
+    #[inline]
+    pub fn keyword_table_for_version(version: &str) -> Vec<String> {
+        ffi::SyntaxToken::keyword_table_for_version(CxxSV::new(version))
+    }
+
+    #[inline]
+    pub fn keyword_kind_for_version(version: &str, text: &str) -> TokenKind {
+        TokenKind::from_id(ffi::SyntaxToken::keyword_kind_for_version(
+            CxxSV::new(version),
+            CxxSV::new(text),
+        ))
+    }
+
+    #[inline]
+    pub fn verilog_2005_keywords() -> Vec<String> {
+        ffi::SyntaxToken::verilog_2005_keywords()
+    }
+
+    #[inline]
+    pub fn directive_text(kind: SyntaxKind) -> String {
+        ffi::SyntaxToken::directive_text(kind.as_u16())
+    }
+
+    #[inline]
+    fn from_raw_ptr(_ptr: *const ffi::SyntaxToken) -> Option<Self> {
+        _ptr.is_null().not().then(|| SyntaxToken { _ptr: unsafe { Pin::new_unchecked(&*_ptr) } })
+    }
+
+    #[inline]
+    pub fn is_missing(&self) -> bool {
+        self._ptr.isMissing()
+    }
+
+    #[inline]
+    pub fn range(&self) -> Option<SourceRange> {
+        SourceRange::from_unique_ptr(self._ptr.range())
+    }
+
+    #[inline]
+    pub fn value_text(&self) -> CxxSV<'_> {
+        self._ptr.valueText()
+    }
+
+    #[inline]
+    pub fn raw_text(&self) -> CxxSV<'_> {
+        self._ptr.rawText()
+    }
+
+    #[inline]
+    pub fn kind(&self) -> TokenKind {
+        TokenKind::from_id(self._ptr.kind())
+    }
+
+    #[inline]
+    pub fn int(&self) -> Option<SVInt> {
+        matches!(self.kind(), TokenKind::INTEGER_LITERAL)
+            .then(|| SVInt { _ptr: self._ptr.intValue() })
+    }
+
+    #[inline]
+    pub fn bits(&self) -> Option<SVLogic> {
+        matches!(self.kind(), TokenKind::UNBASED_UNSIZED_LITERAL)
+            .then(|| SVLogic { _ptr: self._ptr.bitValue() })
+    }
+
+    #[inline]
+    pub fn real(&self) -> Option<f64> {
+        matches!(self.kind(), TokenKind::REAL_LITERAL | TokenKind::TIME_LITERAL)
+            .then(|| self._ptr.realValue())
+    }
+
+    #[inline]
+    pub fn base(&self) -> Option<LiteralBase> {
+        matches!(self.kind(), TokenKind::INTEGER_BASE)
+            .then(|| unsafe { std::mem::transmute::<u8, LiteralBase>(self._ptr.base()) })
+    }
+
+    #[inline]
+    pub fn time_unit(&self) -> Option<TimeUnit> {
+        matches!(self.kind(), TokenKind::TIME_LITERAL)
+            .then(|| unsafe { std::mem::transmute::<u8, TimeUnit>(self._ptr.unit()) })
+    }
+
+    #[inline]
+    pub fn trivia_count(&self) -> usize {
+        self._ptr.trivia_count()
+    }
+
+    #[inline]
+    pub fn trivia_at(&self, idx: usize) -> Option<SyntaxTrivia<'a>> {
+        SyntaxTrivia::from_raw_ptr(self._ptr.trivia(idx))
+    }
+
+    #[inline]
+    pub fn trivias(&self) -> impl ChildrenIter<SyntaxTrivia<'a>> + use<'a> {
+        SyntaxTriviaIter { tok: *self, idx: 0, total: self.trivia_count() }
+    }
+
+    #[inline]
+    pub fn trivias_with_loc(
+        &self,
+    ) -> impl ChildrenIter<(SyntaxTriviaLoc, SyntaxTrivia<'a>)> + use<'a> {
+        let Some(range) = self.range() else {
+            return Either::Left(iter::empty());
+        };
+        let mut cursor_buffer_id = range.start_buffer_id();
+        let mut cursor_offset = range.start();
+        let mut locs = Vec::with_capacity(self.trivia_count());
+
+        for trivia in self.trivias().rev() {
+            let len = trivia.get_raw_text().as_bytes().len();
+
+            let loc = if let Some(location) = trivia.explicit_location() {
+                let start = location.offset().unwrap();
+                SyntaxTriviaLoc {
+                    buffer_id: location.buffer_id().unwrap(),
+                    start,
+                    end: start + len,
+                }
+            } else {
+                let end = cursor_offset;
+                let start = end - len;
+                SyntaxTriviaLoc { buffer_id: cursor_buffer_id, start, end }
+            };
+
+            cursor_buffer_id = loc.buffer_id;
+            cursor_offset = loc.start;
+            locs.push((loc, trivia));
+        }
+
+        Either::Right(locs.into_iter().rev())
+    }
+}
+
+impl fmt::Debug for SyntaxToken<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SyntaxToken")
+            .field("kind", &self.kind())
+            .field("range", &self.range())
+            .field("value_text", &self.value_text())
+            .finish()
+    }
+}
+
+impl PartialEq for SyntaxToken<'_> {
+    fn eq(&self, other: &Self) -> bool {
+        self.kind() == other.kind() && self.range() == other.range()
+    }
+}
+
+impl Eq for SyntaxToken<'_> {}
+
+#[derive(Debug, Clone)]
+pub struct SyntaxTriviaIter<'a> {
+    tok: SyntaxToken<'a>,
+    idx: usize,
+    total: usize,
+}
+
+impl<'a> Iterator for SyntaxTriviaIter<'a> {
+    type Item = SyntaxTrivia<'a>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        if self.idx < self.total {
+            let trivia = self.tok.trivia_at(self.idx).unwrap();
+            self.idx += 1;
+            Some(trivia)
+        } else {
+            None
+        }
+    }
+}
+
+impl<'a> DoubleEndedIterator for SyntaxTriviaIter<'a> {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        if self.idx < self.total {
+            self.total -= 1;
+            let trivia = self.tok.trivia_at(self.total).unwrap();
+            Some(trivia)
+        } else {
+            None
+        }
+    }
+}
+
+impl<'a> ExactSizeIterator for SyntaxTriviaIter<'a> {
+    fn len(&self) -> usize {
+        self.total - self.idx
+    }
+}
+
+impl<'a> SyntaxNode<'a> {
+    #[inline]
+    fn from_raw_ptr(_ptr: *const ffi::SyntaxNode) -> Option<Self> {
+        _ptr.is_null().not().then(|| SyntaxNode { _ptr: unsafe { Pin::new_unchecked(&*_ptr) } })
+    }
+
+    #[inline]
+    pub fn walk(&self) -> SyntaxCursor<'a> {
+        SyntaxCursor::new(*self)
+    }
+
+    #[inline]
+    pub fn range(&self) -> Option<SourceRange> {
+        SourceRange::from_unique_ptr(self._ptr.range())
+    }
+
+    #[inline]
+    pub fn range_with_context(&self, context: SyntaxNode<'a>) -> Option<SourceRange> {
+        let node = self._ptr.as_ref().get_ref();
+        let context = context._ptr.as_ref().get_ref();
+        SourceRange::from_unique_ptr(ffi::SyntaxNode::rangeWithContext(node, context))
+    }
+
+    #[inline]
+    pub fn child_node(&self, idx: usize) -> Option<SyntaxNode<'a>> {
+        SyntaxNode::from_raw_ptr(self._ptr.childNode(idx))
+    }
+
+    // not-null
+    #[inline]
+    pub fn child_token(&self, idx: usize) -> Option<SyntaxToken<'a>> {
+        SyntaxToken::from_raw_ptr(self._ptr.childToken(idx))
+            .filter(|tok| tok.kind() != TokenKind::UNKNOWN)
+    }
+
+    #[inline]
+    pub fn child_count(&self) -> usize {
+        self._ptr.getChildCount()
+    }
+
+    #[inline]
+    pub fn kind(&self) -> SyntaxKind {
+        SyntaxKind::from_id(self._ptr.kind())
+    }
+
+    #[inline]
+    pub fn parent(&self) -> Option<SyntaxNode<'a>> {
+        SyntaxNode::from_raw_ptr(self._ptr.parent())
+    }
+
+    #[inline]
+    pub fn child(&self, idx: usize) -> Option<SyntaxElement<'a>> {
+        // TODO: we have to visit twice to get the child, this is not efficient
+        if idx >= self.child_count() {
+            None
+        } else if let Some(node) = self.child_node(idx) {
+            Some(SyntaxElement::Node(node))
+        } else {
+            self.child_token(idx)
+                .map(|tok| SyntaxElement::Token(SyntaxTokenWithParent { parent: *self, tok }))
+        }
+    }
+
+    #[inline]
+    pub fn children_with_idx(&self) -> SyntaxIdxChildren<'a> {
+        SyntaxIdxChildren::new(*self)
+    }
+
+    #[inline]
+    pub fn children(&self) -> SyntaxChildren<'a> {
+        SyntaxChildren::new(*self)
+    }
+
+    #[inline]
+    pub fn elem_preorder(&self) -> SyntaxElemPreorder<'a> {
+        SyntaxElemPreorder::new(*self)
+    }
+
+    #[inline]
+    pub fn node_preorder(&self) -> SyntaxNodePreorder<'a> {
+        SyntaxNodePreorder::new(*self)
+    }
+
+    #[inline]
+    pub fn first_token(&self) -> Option<SyntaxTokenWithParent<'a>> {
+        let mut cursor = self.walk();
+
+        while cursor.to_tok_with_parent().is_none() {
+            if cursor.goto_first_child() {
+                continue;
+            }
+
+            while !cursor.goto_next_sibling() {
+                if !cursor.goto_parent() {
+                    unreachable!("Tree has no tokens");
+                }
+            }
+        }
+
+        cursor.to_tok_with_parent()
+    }
+}
+
+impl fmt::Debug for SyntaxNode<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("SyntaxNode")
+            .field("kind", &self.kind())
+            .field("range", &self.range())
+            .field("child_count", &self.child_count())
+            .finish()
+    }
+}
+
+impl PartialEq for SyntaxNode<'_> {
+    #[inline]
+    fn eq(&self, other: &Self) -> bool {
+        // Just compare pointer
+        std::ptr::eq(Pin::as_ref(&self._ptr).get_ref(), Pin::as_ref(&other._ptr).get_ref())
+    }
+}
+
+impl Eq for SyntaxNode<'_> {}
+
+impl hash::Hash for SyntaxNode<'_> {
+    #[inline]
+    fn hash<H: hash::Hasher>(&self, state: &mut H) {
+        let ptr = Pin::as_ref(&self._ptr).get_ref() as *const ffi::SyntaxNode;
+        ptr.hash(state)
+    }
+}
+
+fn raw_include_buffers(options: &SyntaxTreeOptions) -> Vec<ffi::RawSourceBuffer> {
+    options
+        .include_buffers
+        .iter()
+        .map(|buffer| ffi::RawSourceBuffer { path: buffer.path.clone(), text: buffer.text.clone() })
+        .collect()
+}
+
+impl SyntaxTree {
+    #[inline]
+    pub fn from_text(text: &str, name: &str, path: &str) -> SyntaxTree {
+        SyntaxTree {
+            _ptr: ffi::SyntaxTree::fromText(CxxSV::new(text), CxxSV::new(name), CxxSV::new(path)),
+        }
+    }
+
+    #[inline]
+    pub fn from_text_with_options(
+        text: &str,
+        name: &str,
+        path: &str,
+        options: &SyntaxTreeOptions,
+    ) -> SyntaxTree {
+        SyntaxTree {
+            _ptr: ffi::SyntaxTree::fromTextWithOptions(
+                CxxSV::new(text),
+                CxxSV::new(name),
+                CxxSV::new(path),
+                options.predefines.clone(),
+                options.include_paths.clone(),
+                raw_include_buffers(options),
+                options.expand_includes,
+            ),
+        }
+    }
+
+    #[inline]
+    pub fn from_text_with_options_and_trace(
+        text: &str,
+        name: &str,
+        path: &str,
+        options: &SyntaxTreeOptions,
+    ) -> SyntaxTreeWithPreprocessorTrace {
+        let tree = SyntaxTree {
+            _ptr: ffi::SyntaxTree::fromTextWithOptionsAndTrace(
+                CxxSV::new(text),
+                CxxSV::new(name),
+                CxxSV::new(path),
+                options.predefines.clone(),
+                options.include_paths.clone(),
+                raw_include_buffers(options),
+                options.expand_includes,
+            ),
+        };
+        let preprocessor_trace =
+            PreprocessorTrace::from_raw(tree._ptr.preprocessorTraceFromParsed());
+        SyntaxTreeWithPreprocessorTrace { tree, preprocessor_trace }
+    }
+
+    #[inline]
+    pub fn from_library_map_text(text: &str, name: &str, path: &str) -> SyntaxTree {
+        SyntaxTree {
+            _ptr: ffi::SyntaxTree::fromLibraryMapText(
+                CxxSV::new(text),
+                CxxSV::new(name),
+                CxxSV::new(path),
+            ),
+        }
+    }
+
+    #[inline]
+    pub fn root(&self) -> Option<SyntaxNode<'_>> {
+        SyntaxNode::from_raw_ptr(self._ptr.root())
+    }
+
+    pub fn diagnostics(&self) -> Vec<SyntaxDiagnostic> {
+        self._ptr.diagnostics().into_iter().map(SyntaxDiagnostic::from_raw).collect()
+    }
+
+    pub fn diagnostics_with_options(&self, warning_options: &[String]) -> Vec<SyntaxDiagnostic> {
+        self._ptr
+            .diagnostics_with_options(warning_options.to_vec())
+            .into_iter()
+            .map(SyntaxDiagnostic::from_raw)
+            .collect()
+    }
+
+    pub fn expected_syntax_at_offset(
+        text: &str,
+        name: &str,
+        path: &str,
+        offset: usize,
+    ) -> Vec<ParserExpectedSyntax> {
+        Self::expected_syntax_at_offset_with_options(
+            text,
+            name,
+            path,
+            offset,
+            &SyntaxTreeOptions::default(),
+        )
+    }
+
+    pub fn expected_syntax_at_offset_with_options(
+        text: &str,
+        name: &str,
+        path: &str,
+        offset: usize,
+        options: &SyntaxTreeOptions,
+    ) -> Vec<ParserExpectedSyntax> {
+        ffi::SyntaxTree::expectedSyntaxAtOffset(
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+            offset,
+            options.predefines.clone(),
+            options.include_paths.clone(),
+            raw_include_buffers(options),
+            options.expand_includes,
+        )
+        .into_iter()
+        .map(ParserExpectedSyntax::from_raw)
+        .collect()
+    }
+
+    pub fn library_map_expected_syntax_at_offset(
+        text: &str,
+        name: &str,
+        path: &str,
+        offset: usize,
+    ) -> Vec<ParserExpectedSyntax> {
+        ffi::SyntaxTree::libraryMapExpectedSyntaxAtOffset(
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+            offset,
+        )
+        .into_iter()
+        .map(ParserExpectedSyntax::from_raw)
+        .collect()
+    }
+
+    pub fn directive_at_offset(
+        text: &str,
+        name: &str,
+        path: &str,
+        offset: usize,
+    ) -> Option<LexedTokenAtOffset> {
+        LexedTokenAtOffset::from_raw(ffi::SyntaxTree::directiveAtOffset(
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+            offset,
+        ))
+    }
+
+    pub fn token_word_at_offset(
+        text: &str,
+        name: &str,
+        path: &str,
+        offset: usize,
+    ) -> Option<LexedTokenAtOffset> {
+        LexedTokenAtOffset::from_raw(ffi::SyntaxTree::tokenWordAtOffset(
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+            offset,
+        ))
+    }
+
+    pub fn buffer_id(&self) -> u32 {
+        self._ptr.buffer_id()
+    }
+}
+
+unsafe impl Send for SyntaxTree {}
+
+unsafe impl Sync for SyntaxTree {}
+
+impl fmt::Debug for SyntaxTree {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("<SyntaxTree>").finish()
+    }
+}
+
+impl PartialEq for SyntaxTree {
+    #[inline]
+    fn eq(&self, other: &Self) -> bool {
+        let a = self._ptr.as_ref().unwrap();
+        let b = other._ptr.as_ref().unwrap();
+        std::ptr::eq(std::ptr::from_ref(a), std::ptr::from_ref(b))
+    }
+}
+
+impl Eq for SyntaxTree {}
+
+#[derive(Debug, Clone, Copy)]
+pub struct SyntaxTokenWithParent<'a> {
+    pub parent: SyntaxNode<'a>,
+    pub tok: SyntaxToken<'a>,
+}
+
+impl SyntaxTokenWithParent<'_> {
+    #[inline]
+    pub fn range(&self) -> Option<SourceRange> {
+        let token = self.tok._ptr.as_ref().get_ref();
+        let context = self.parent._ptr.as_ref().get_ref();
+        SourceRange::from_unique_ptr(ffi::SyntaxToken::rangeWithContext(token, context))
+    }
+
+    #[inline]
+    pub fn preprocessor_trace_provenance(&self) -> PreprocessorTraceTokenProvenance {
+        let token = self.tok._ptr.as_ref().get_ref();
+        let context = self.parent._ptr.as_ref().get_ref();
+        PreprocessorTraceEmittedToken::from_raw(
+            ffi::SyntaxToken::preprocessorTraceProvenanceWithContext(token, context),
+        )
+        .provenance
+    }
+}
+
+impl<'a> std::ops::Deref for SyntaxTokenWithParent<'a> {
+    type Target = SyntaxToken<'a>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.tok
+    }
+}
+
+impl PartialEq for SyntaxTokenWithParent<'_> {
+    fn eq(&self, other: &Self) -> bool {
+        self.parent == other.parent && self.tok == other.tok
+    }
+}
+
+impl Eq for SyntaxTokenWithParent<'_> {}
+
+#[derive(Debug, Clone, Copy)]
+pub enum SyntaxElement<'a> {
+    Node(SyntaxNode<'a>),
+    Token(SyntaxTokenWithParent<'a>),
+}
+
+impl<'a> SyntaxElement<'a> {
+    pub fn from_node(node: SyntaxNode) -> SyntaxElement {
+        SyntaxElement::Node(node)
+    }
+
+    pub fn from_token<'b>(tok_with_parent: SyntaxTokenWithParent<'b>) -> SyntaxElement<'b> {
+        SyntaxElement::Token(tok_with_parent)
+    }
+
+    pub fn as_node(&self) -> Option<SyntaxNode<'a>> {
+        match self {
+            SyntaxElement::Node(node) => Some(*node),
+            SyntaxElement::Token(_) => None,
+        }
+    }
+
+    pub fn as_tok_with_parent(&self) -> Option<SyntaxTokenWithParent<'a>> {
+        match self {
+            SyntaxElement::Token(tok_with_parent) => Some(*tok_with_parent),
+            SyntaxElement::Node(_) => None,
+        }
+    }
+
+    pub fn as_token(&self) -> Option<SyntaxToken<'a>> {
+        match self {
+            SyntaxElement::Token(tok_with_parent) => Some(tok_with_parent.tok),
+            SyntaxElement::Node(_) => None,
+        }
+    }
+
+    pub fn child_count(&self) -> usize {
+        match self {
+            SyntaxElement::Node(node) => node.child_count(),
+            SyntaxElement::Token { .. } => 0,
+        }
+    }
+
+    pub fn child(&self, idx: usize) -> Option<SyntaxElement<'a>> {
+        match self {
+            SyntaxElement::Node(node) => node.child(idx),
+            SyntaxElement::Token { .. } => None,
+        }
+    }
+
+    pub fn range(&self) -> Option<SourceRange> {
+        match self {
+            SyntaxElement::Node(node) => node.range(),
+            SyntaxElement::Token(tok) => tok.range(),
+        }
+    }
+
+    pub fn parent(&self) -> Option<SyntaxNode<'a>> {
+        match self {
+            SyntaxElement::Node(node) => node.parent(),
+            SyntaxElement::Token(tok) => Some(tok.parent),
+        }
+    }
+
+    pub fn kind(&self) -> SyntaxElementKind {
+        match self {
+            SyntaxElement::Node(node) => SyntaxElementKind::Node(node.kind()),
+            SyntaxElement::Token(tok) => SyntaxElementKind::Token(tok.kind()),
+        }
+    }
+
+    pub fn children_with_idx(
+        &self,
+    ) -> Either<SyntaxIdxChildren<'a>, iter::Empty<(usize, SyntaxElement<'a>)>> {
+        match self {
+            SyntaxElement::Node(node) => Either::Left(node.children_with_idx()),
+            SyntaxElement::Token(_) => Either::Right(iter::empty()),
+        }
+    }
+
+    pub fn children(&self) -> Either<SyntaxChildren<'a>, iter::Empty<SyntaxElement<'a>>> {
+        match self {
+            SyntaxElement::Node(node) => Either::Left(node.children()),
+            SyntaxElement::Token(_) => Either::Right(iter::empty()),
+        }
+    }
+}
+
+impl<'a> From<SyntaxNode<'a>> for SyntaxElement<'a> {
+    fn from(node: SyntaxNode<'a>) -> SyntaxElement<'a> {
+        SyntaxElement::Node(node)
+    }
+}
+
+impl<'a> From<SyntaxTokenWithParent<'a>> for SyntaxElement<'a> {
+    fn from(tok_with_parent: SyntaxTokenWithParent<'a>) -> SyntaxElement<'a> {
+        SyntaxElement::Token(tok_with_parent)
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum SyntaxElementKind {
+    Node(SyntaxKind),
+    Token(TokenKind),
+}
+
+impl SyntaxElementKind {
+    pub fn is_list(&self) -> bool {
+        match self {
+            SyntaxElementKind::Node(kind) => kind.is_list(),
+            SyntaxElementKind::Token(_) => false,
+        }
+    }
+}
+
+impl From<SyntaxKind> for SyntaxElementKind {
+    fn from(kind: SyntaxKind) -> SyntaxElementKind {
+        SyntaxElementKind::Node(kind)
+    }
+}
+
+impl From<TokenKind> for SyntaxElementKind {
+    fn from(kind: TokenKind) -> SyntaxElementKind {
+        SyntaxElementKind::Token(kind)
+    }
+}
+
+pub struct Compilation {
+    _ptr: UniquePtr<ffi::Compilation>,
+}
+
+impl Default for Compilation {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Compilation {
+    pub fn new() -> Self {
+        Compilation { _ptr: ffi::Compilation::new() }
+    }
+
+    pub fn new_with_top_modules(top_modules: &[String]) -> Self {
+        Compilation { _ptr: ffi::Compilation::new_with_top_modules(top_modules.to_vec()) }
+    }
+
+    pub fn system_function_names() -> Vec<String> {
+        ffi::Compilation::system_function_names()
+    }
+
+    pub fn system_task_names() -> Vec<String> {
+        ffi::Compilation::system_task_names()
+    }
+
+    pub fn add_syntax_tree(&mut self, tree: SyntaxTree) {
+        ffi::Compilation::add_syntax_tree(self._ptr.as_mut().unwrap(), tree._ptr);
+    }
+
+    pub fn add_syntax_tree_from_text(
+        &mut self,
+        text: &str,
+        name: &str,
+        path: &str,
+        options: &SyntaxTreeOptions,
+    ) -> SyntaxTreeBufferIds {
+        SyntaxTreeBufferIds::from_raw(ffi::Compilation::add_syntax_tree_from_text(
+            self._ptr.as_mut().unwrap(),
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+            options.predefines.clone(),
+            options.include_paths.clone(),
+            options
+                .include_buffers
+                .iter()
+                .map(|buffer| ffi::RawSourceBuffer {
+                    path: buffer.path.clone(),
+                    text: buffer.text.clone(),
+                })
+                .collect(),
+            options.expand_includes,
+        ))
+    }
+
+    pub fn add_library_map_syntax_tree_from_text(
+        &mut self,
+        text: &str,
+        name: &str,
+        path: &str,
+    ) -> SyntaxTreeBufferIds {
+        SyntaxTreeBufferIds::from_raw(ffi::Compilation::add_library_map_syntax_tree_from_text(
+            self._ptr.as_mut().unwrap(),
+            CxxSV::new(text),
+            CxxSV::new(name),
+            CxxSV::new(path),
+        ))
+    }
+
+    pub fn semantic_diagnostics(&self) -> Vec<SyntaxDiagnostic> {
+        self._ptr.semantic_diagnostics().into_iter().map(SyntaxDiagnostic::from_raw).collect()
+    }
+
+    pub fn parse_diagnostics_with_options(
+        &self,
+        warning_options: &[String],
+    ) -> Vec<SyntaxDiagnostic> {
+        self._ptr
+            .parse_diagnostics_with_options(warning_options.to_vec())
+            .into_iter()
+            .map(SyntaxDiagnostic::from_raw)
+            .collect()
+    }
+
+    pub fn semantic_diagnostics_with_options(
+        &self,
+        warning_options: &[String],
+    ) -> Vec<SyntaxDiagnostic> {
+        self._ptr
+            .semantic_diagnostics_with_options(warning_options.to_vec())
+            .into_iter()
+            .map(SyntaxDiagnostic::from_raw)
+            .collect()
+    }
+}
